Add convenience methods for executing commands
This follows a similar approach to the Path
helper and extracts methods for each call to
Exec so that Provider and its specs are much
simpler.

I also added stubs for Exec to spec_helper.rb
so that we don't execute commands in the
tests, except where they're explicitly
unstubbed to unit test .execute and .capture.