Run chef with additional permissions
In order to install homebrew, the script needs to
be run as the current user, but it also requires
sudo rights to change the ownership of
/usr/local/bin amongst other things.

Chef is being run in a subprocess and as a result,
there’s no TTY associated with the shell. This
means Homebrew can’t ask for a password when it
runs sudo and the process fails.

This workaround temporarily grants the user the
right to run sudo commands without a password for
the duration of the chef run. It revokes them at
the end, even if chef fails.

There’s some discussion about this here:
https://github.com/chef-cookbooks/homebrew/issues/105