Download and install Dropbox with chef
This one’s a bit tricky because Dropbox doesn’t
support logging into the desktop app via command
line. To workaround this, we open the UI and wait
for authentication. A good heuristic for whether
they’ve done this is if the ~/.dropbox/info.json
file exists.

Finally, we wait until files needed for the rest
of the chef run are synced. I keep my RSA and GPG
keys in Dropbox, which is controversial, but these
private keys have strong passwords set on them and
a high number of KDF rounds.