Install iterm and load its config from a directory
iterm has a feature that allows configuration to
be read from a directory. This creates a directory
containing a config xml file and uses the
'defaults' command to set iterm's preferences to
point to it.

The process of setting iterm's preferences is
abstracted by a Pref class that will likely be
useful for configuring other apps.

The xml config file is a minimal version. I
plan to configure this in the next commit.