Improve Vim support for Rust
- cargo is now the default syntactic checker
- inherit RUSTFLAGS from profile
- add leader bindings for build/run/test

This also adds ‘leader-a’ which helps address an
issue with ‘cargo check’. This doesn’t reuse the
same target as ‘cargo test’ so dependencies have
to be built twice when adding a crate.

If we don’t run ‘cargo check’ manually, then it
means syntastic runs it when the next rust file is
saved, which can take a while and it doesn’t give
any output for the crates it’s installing.