Try to fix a long-standing problem I’ve been having with ssh
My ssh key is derived from my private gpg key,
which is stored on my Yubikey. When I use ssh, for
example, when git pull-ing, my Yubikey needs to be
‘unlocked’ so that the gpg-agent can read my gpg
key.

By default, ssh doesn’t tell the gpg-agent which
tty it has been called from. When gpg-agent
starts, it remembers which tty was set at that
time and uses that for future pinentry requests.

However, if the tty is different, e.g. by opening
a new terminal tab, the pinentry request will
appear in the wrong place and the ssh command will
hang without any feedback.

I’d previously been using a custom gpg-reset
command to get around this problem which tells
gpg-agent to use the current tty for pinentry
requests but this is awkward.

This comment tries to fix this by adding a match
rule into ~/.ssh/config that tells gpg-agent to
use the current tty immediately before any ssh
command is run. This seems to require a newer
version of ssh so I’m using the openssh package.

https://unix.stackexchange.com/questions/280879/how-to-get-pinentry-curses-to-start-on-the-correct-tty#answer-499133