Auto-complete command and option names
Try to infer whether the user is trying to complete a command
or option name based on how many arguments are provided. On an
exact match of a command, complete option names for that command.

If an option takes no arguments, then complete then offer the
full list of options for that command again, otherwise print nothing.
I intend to hand-over control of the completion list to the option.